profiler-hub: split reader_test.cpp into six per-area test files
Splits tests/unit/reader_test.cpp (4280 lines, 128 TEST_F, 22 fixture
classes) into reader_core_test.cpp, reader_flows_test.cpp,
reader_pmc_test.cpp, reader_memory_test.cpp, reader_summary_test.cpp,
and reader_track_metadata_test.cpp, plus a new shared
reader_test_fixture.hpp header holding the base reader_test fixture and
11 helpers used by 2+ files. Pure reorganization: no test added,
removed, renamed, reordered, or edited in body. UNIT_TEST_SOURCES
updated accordingly; no other part of CMakeLists.txt touched.
